Add swagger docs and schedule endpoint
