+++ package.json
@@ -15,7 +15,8 @@
     "cors": "^2.8.5",
     "dotenv": "^16.5.0",
     "express": "^5.1.0",
-    "natural": "^8.1.0"
+    "natural": "^8.1.0",
+    "swagger-ui-express": "^4.7.1"
   },
   "engines": {
     "node": "18.x"
